+++ src/Exception/InvalidServiceException.php
@@ -9,7 +9,7 @@
 
 namespace Zend\ServiceManager\Exception;
 
-use RuntimeException;
+use RuntimeException as SplRuntimeException;
 
 /**
  * InvalidServiceException
@@ -17,6 +17,6 @@ use RuntimeException;
  * This exception is triggered by plugin managers when the created object does not match
  * the plugin manager's conditions
  */
-class InvalidServiceException extends RuntimeException implements ExceptionInterface
+class InvalidServiceException extends SplRuntimeException implements ExceptionInterface
 {
 }
